% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_site}
\alias{build_site}
\title{Build Site}
\usage{
build_site()
}
\description{
This script will build the RMarkdown site stored in the docs directory.
Preferred over simply running \code{rmarkdown::render_site} to avoid regenerating
Rmds that are time-consuming to render or depend on local data that may or
may not be currently available.
}
\details{
Will use a list of Rmds to skip rather than a list of Rmds to build, so that
any new Rmds will be automatically be included in build until specified
otherwise.
}
